change chat id to user id

+++ tg/dto/sticker.go
@@ -2,14 +2,14 @@ package dto
 
 // https://core.telegram.org/bots/api#addstickertoset
 type AddStickerToSetRequest struct {
-	UserID  int          `json:"user_id"`
+	UserID  int64        `json:"user_id"`
 	Name    string       `json:"name"`
 	Sticker InputSticker `json:"sticker"`
 }
 
 // https://core.telegram.org/bots/api#createnewstickerset
 type CreateNewStickerSet struct {
-	UserID   int            `json:"user_id"`
+	UserID   int64          `json:"user_id"`
 	Name     string         `json:"name"`
 	Title    string         `json:"title"`
 	Stickers []InputSticker `json:"stickers"`
